Update OnMove decimal round.

+++ app/src/main/java/com/example/ex4/JoystickListener.java
@@ -2,5 +2,5 @@ package com.example.ex4;
 
 // Listen to joystick movement.
 public interface JoystickListener {
-    void OnMoved(int aileron, int elevator);
+    void OnMoved(double aileron, double elevator);
 }
